# swagger/banking.yml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API
    url: https://virtserver.swaggerhub.com/684006_1/BankingAPI/1.0.0
info:
  version: "1.0.0"
  title: banking-api
  description: The API for the Banking Application by Group 02, IT2A
...
paths:
...
  /accounts:
    get:
      tags:
        - Accounts
        - Employee
        - Client
      security:
        - bearerAuth: []
      description: >
        Returns the list of all accounts belonging to user.<br>
        For **client**, only his own accounts are visible to him.
      operationId: getAccounts
      responses:
        "200":
          description: A complete list of all accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountForClientResponse"
                example:
                  - id: 1
                    kind: current
                    iban: "NL00INHL1234567890"
                    amount: 420.69
                  - id: 2
                    kind: savings
                    iban: "NL00INHL1234567891"
                    amount: 2137.12
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags:
        - Accounts
        - Employee
      security:
        - bearerAuth: []
      description: >
        Creates a new account for "user id" user.
      operationId: createAccount
      parameters:
        - name: user_id
          required: true
          in: query
          description: User ID for which the account is created.
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountRequest"
      responses:
        "200":
          description: "Account created!"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountForClientResponse"
              example:
                id: 1
                iban: "NL00INHL1234567890"
                kind: current
                amount: 0.00
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /accounts/{account_id}:
    get:
      tags:
        - Accounts
        - Employee
        - Client
      security:
        - bearerAuth: []
      description: >
        Returns details of specified account.<br><br>
        _Note: **Client** can only see his own user details._
      operationId: getAccount
      parameters:
        - name: account_id
          in: path
          description: Account ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Account details"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/UserResponse"
                example:
                  id: 1
                  kind: current
                  iban: "NL00INHL1234567890"
                  amount: 100.00
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /transactions:
    post:
      tags:
        - Transaction
      description: >
        Transfers money from one user to another.
      operationId: transfer
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  $ref: "#/components/schemas/TransactionRequest"
      responses:
        "201":
          description: "Transaction successful!"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/TransactionResponse"
                example:
                  id: 1
                  senderiban: "NL00INHL1234567890"
                  receiveriban: "NL00INHL0987654321"
                  amount: 50.00
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags:
        - Transaction
      description: >
        Returns all transactions.
      operationId: getAllTransfers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Returns all transfers."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /transactions/deposit:
    post:
      tags:
        - Employee
        - Client
      description: >
        Deposit money to a account.
      operationId: deposit
      parameters:
        - name: atm_id
          in: query
          description: ATM ID
          required: true
          schema:
            type: integer
            format: int64
        - name: user_id
          in: query
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
        - name: account_id
          in: query
          description: Account ID
          required: true
          schema:
            type: integer
            format: int64
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionDepositRequest"
            example:
              receiver_iban: "NL00INHL1234567890"
              amount: 100.00
      responses:
        "200":
          description: "Deposit successful!"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/TransactionResponse"
                example:
                  id: 1
                  sender_iban: "NL00INHL2234567890"
                  receiver_iban: "NL00INHL1234567890"
                  amount: 150.00
                  timestamp: "2021-06-01T12:00:00.000Z"
                  description: "Deposit"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/InternalServerError"
  /transactions/withdraw:
    post:
      tags:
        - Employee
        - Client
      description: >
        Withdraw money from a account.
      operationId: withdraw
      parameters:
        - name: atm_id
          in: query
          description: ATM ID
          required: true
          schema:
            type: integer
            format: int64
        - name: user_id
          in: query
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
        - name: account_id
          in: query
          description: Account ID
          required: true
          schema:
            type: integer
            format: int64
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionWithdrawRequest"
            example:
              sender_iban: "NL00INHL2234567890"
              amount: 100.00
      responses:
        "200":
          description: "Withdraw successful!"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/TransactionResponse"
                example:
                  id: 1
                  sender_iban: "NL00INHL2234567890"
                  receiver_iban: "NL00INHL1234567890"
                  amount: 50.00
                  timestamp: "2021-06-01T12:00:00.000Z"
                  description: "Withdraw"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        bsn:
          type: string
          pattern: "[0-9]{8,9}"
        phone_number:
          type: string
          pattern: "[0-9]{10,10}"
        birth_date:
          type: string
          format: date
          pattern: "[0-9]{4}-[0-9]{2}-[0-9]{2}"
        total_balance:
          type: number
          format: double
        role:
          type: string
          format: string
          enum: [admin, employee, client]
        current_account:
          $ref: "#/components/schemas/AccountForClientResponse"
        savings_account:
          $ref: "#/components/schemas/AccountForClientResponse"
        active:
          type: boolean
      description: ok
...
    AccountRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: ["current", "saving"]
        iban:
          type: string
          pattern: "[A-Z]{2,2}[0-9]{2,2}[A-Z]{4,4}[0-9]{10,10}"
        balance:
          type: number
          format: double
    AccountForClientResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        iban:
          type: string
          pattern: "[A-Z]{2,2}[0-9]{2,2}[A-Z]{4,4}[0-9]{10,10}"
        type:
          type: string
          enum: ["CURRENT", "SAVING"]
        currency_type:
            type: string
            enum: [ "EURO" ]
        balance:
          type: number
          format: double
    TransactionRequest:
      type: object
      properties:
        sender_iban:
          type: string
          pattern: "[A-Z]{2,2}[0-9]{2,2}[A-Z]{4,4}[0-9]{10,10}"
        receiver_iban:
          type: string
          pattern: "[A-Z]{2,2}[0-9]{2,2}[A-Z]{4,4}[0-9]{10,10}"
        amount:
          type: number
          format: double
        description:
          type: string
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sender_iban:
          type: string
          pattern: "[A-Z]{2,2}[0-9]{2,2}[A-Z]{4,4}[0-9]{10,10}"
          nullable: true
        receiver_iban:
          type: string
          pattern: "[A-Z]{2,2}[0-9]{2,2}[A-Z]{4,4}[0-9]{10,10}"
          nullable: true
        amount:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
        description:
          type: string
    TransactionDepositRequest:
      type: object
      properties:
        receiver_iban:
          type: string
          pattern: "[A-Z]{2,2}[0-9]{2,2}[A-Z]{4,4}[0-9]{10,10}"
        amount:
          type: number
          format: double
        description:
          type: string
    TransactionWithdrawRequest:
      type: object
      properties:
        sender_iban:
          type: string
          pattern: "[A-Z]{2,2}[0-9]{2,2}[A-Z]{4,4}[0-9]{10,10}"
        amount:
          type: number
          format: double
        description:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: "Authorization information is missing or is invalid."
...
    BadRequest:
      description: "Request body is incorrect."
    NotFound:
      description: "Element does not exist."
...
    InternalServerError:
      description: "Internal server error."